Fix mirror mode when samples fall at integers in RP
In a reflected range of mirror mode we need to snap the opposite way at integers to create a consistent sampling across the logical infinite image.

This change is necessary to make it so that bilerp behaves the same as nearest when sampling 1:1 with no offset in mirror tile mode. In that case bilerp's +/-.5 offset samples fall exactly at integers. Currently SkImageShader detects when it is "safe" to downgrade bilerp sampling to nearest. However, this assumes SkImageShader can know a "total matrix" between device and image space. This assumption is incompatible with general SkSL shaders.

Bug: skia:13752
Change-Id: I05280e582334b4da13d1ee8bdb8d059107cb0c65
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/615319
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>