[bazel] Implement layering_check in our custom toolchain
Important thing we learned about Bazel sandboxing:
    "Technically, the compilation of a .cc file may transitively
     include any header file in the hdrs or srcs in any cc_library
     in the transitive deps closure." - Bazel docs [1]

The sandbox makes all these headers available; public
headers can and do include private headers
(skia_public certainly does this). Clients are only *supposed*
to use the ones from the public hdrs, but Bazel's sandbox
cannot enforce this.

In order to enforce this, Bazel uses Clang's module map
files (not to be confused with C++20 modules) to enforce
this during compilation. [2]

Bazel generates .cppmap files for each cc_library it builds [5].
For example:

    module "//experimental/bazel_test/core:core" {
        export *
        textual header "../../../../../../experimental/bazel_test/core/core.h"
        private textual header "../../../../../../experimental/bazel_test/core/core_priv.h"
        use "//experimental/bazel_test/base:base"
        use "//experimental/bazel_test/base:base_priv"
        use "crosstool"
    }
    extern module "//experimental/bazel_test/base:base" "relative/path/to/base/base.cppmap"
    extern module "//experimental/bazel_test/base:base_priv" "relative/path/to/base//base_priv.cppmap"
    extern module "crosstool" "relative/path/to/external/clang_mac/toolchain_system_headers.modulemap"

The "crosstool" module name is what Bazel refers to the system
headers as, which corresponds to the generated .cppmap [3] [4].

Then, Clang can use this information to determine if an include
is allowable (i.e. it is public) and specified (i.e. not only
from a transitive include).

I tested this by hand on Linux and Mac systems. I plan to add
a CI job in a follow-up CL, using the files and defines in
experimental/bazel_test/, which is a small scale example
of having public and private header files that Skia wants to
use but not necessarily expose.

Currently, the //:skia_public build does not pass this layering_check,
so it is disabled by default. This is just because of how it
was designed (intentionally having third_party deps flow up
via cc_library instead of lists).

As seen in //experimental/bazel_test/client/BUILD.bazel,
if we want to enforce layering checks on headers, we'll need
to do a similar thing to enforcing IWYU on headers - generate
a .h.cpp file from them and compile them (with slightly
different copts). G3 already compiles headers on their own,
which is one fewer clients to worry about. [6]

Suggested review order:
 - toolchain/clang_layering_check.bzl to see the extra Clang
   compilation flags added. Note the module_map_file is
   something from //toolchain/BUILD.bazel
 - toolchain/download_linux_amd64_toolchain.bzl to see
   the system header file map is created at toolchain
   install time, using the same script that the default
   Bazel toolchain uses.
 - experimental/bazel_test/... to see a test setup
   (which mirrors our design sketch of a "Skia Base" API)
 - bazel/buildrc to add a config to disable sandboxing
   *and* layering_check since the latter requires the former.
 - All other files

[1] https://bazel.build/reference/be/c-cpp#hdrs
[2] https://clang.llvm.org/docs/Modules.html
[3] https://github.com/bazelbuild/bazel/blob/8f5b626acea0086be8a314d5efbf6bc6d3473cd2/tools/cpp/generate_system_module_map.sh#L18
[4] https://maskray.me/blog/2022-09-25-layering-check-with-clang
[5] https://github.com/bazelbuild/bazel/blob/4e940a73151e2e68f8eb8725a3c743ea30013691/src/main/java/com/google/devtools/build/lib/rules/cpp/CppModuleMapAction.java#L42
[6] go/cpp-features#parse-headers

Change-Id: I910fd8240a87d8fda2957a904ccb8e79fa61b7b9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/614797
Reviewed-by: Herb Derby <herb@google.com>