Add mask_off_loop_mask Raster Pipeline op.
This disables execution on any running lanes via the loop mask, which
is what the `break` statement does. It is analogous to
`update_return_mask`, which is what the `return` statement does for
the return-mask. So much so, in fact, that I've renamed
`update_return_mask` to `mask_off_return_mask` for consistency.

Change-Id: I9b479cde8977da7e94082916c8041cbb6b620b13
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/616556
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>