Add continue statement to RP codegen.
This uses the same basic pattern as the VM code generator. A slot
is set aside to be used as a continue mask, and initialized to zero
at the top of every loop iteration. If a continue is encountered,
the continue-mask is enabled on those lanes, then execution is
masked off on those lanes. At the end of the loop body, the
continue mask is re-enabled in the loop body.

Right now this is not really optimized--loops which do not contain
a continue still incur a zero-fill and a bitwise-or for the
continue mask on each iteration. Also, the continue op itself is
not optimized (it's built out of existing ops). For now, I'm not
too worried about it, but it definitely gives us room for
improvement when we start looking at performance in earnest.

Change-Id: Ia61aad5f7614305ebd286d8d4ae6642e10732b92
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/616562
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>