Remove dependency of SkPaint -> SkPaintPriv
If we want SkPaint to be a part of "core", then we need
to trim down the .cpp file a bit. This moves all
SkPaintPriv:: implementations to SkPaintPriv.cpp
(Which were just Flatten and Unflatten).

ComputeResScaleForStroking seems to be more related to
Matrices and Paths than Paint, so I moved it to
SkMatrixPriv (and you'll see in a follow-up CL I'm attempting
the ambitious removal of getFillPath)

Change-Id: I14b3d6fb2c69bdeacee998c0d0bc436c59096300
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/616818
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Herb Derby <herb@google.com>