Simplify ternaries that don't have any side effects.
While working on && and ||, it struck me that we could avoid some
of the condition-mask juggling in pushTernaryExpression in cases
where there are no side effects in the true- and false-expressions.
This is likely to be a common case, so it's worth adding code to
support it.

Change-Id: I89d12b235b04f52cdd9dfa370cdc045e801d1338
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/616916
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>