Fix miter issue with AAConvexTessellator.
For normal quadratic and cubic curves we set the final CurveState
to be Indeterminate, and let neighboring bisector angles determine
what that state should be. But for lines we set the CurveState
to Sharp -- which should also be the case for quads and cubics which
are effectively lines.

Bug: skia:13909
Change-Id: Ic429d7099c66a734ac57fedaec5b2ff68bac8abf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/620846
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>