Introduce SkPathEnums.h
This also has some misc cleanups of SkPathRef including
removal of unused methods and moving an SkPath implementation
to SkPath.cpp.

As we modularize things into base and core, we want
SkPathRef to be in core, but not used outside of core.
Thus, we need to any dependencies from stuff outside of
core on SkPathRef (e.g. effects/gpu). This is achieved
by moving the enums those things reference to SkPathEnums.h
(which will eventually be in base).

I moved kPathRefGenIDBitCnt from SkPathPriv.h to SkPathRef.cpp
because it was only used in SkPath.cpp and SkPathRef.cpp
(and we might want to decouple SkPath.cpp from SkPathPriv.h
anyway) and now only used directly by SkPathRef.cpp.

Change-Id: Iccc1d0ba822ad374205ec798ef2f44cc42f9e107
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/621196
Reviewed-by: Michael Ludwig <michaelludwig@google.com>