[direct3d] Fix return value from boolean to pointer type
The GrD3DTexture::MakeAliasingTexture returns a pointer
of an object type. In case of an failure, the return
value should be 'nullptr' instead of 'false' because the
later one is a boolean literal. This fixes the following
compiler error with gcc toolchain

GrD3DTexture.cpp:126:16: error: could not convert 'false' from 'bool' to 'sk_sp<GrD3DTexture>'

Change-Id: I023e5cc86f321a5e2a1dccaa19462e2612678ba8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/621916
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>