Fix recovery procedure when talking to Android devices.
It turns out that `android wait-for-device` after rebooting
a device only waits for the kernel to be up and running, and not
other services, like sdcardfs, which will lead to things like
the /sdcard/ doesn't exist, or is read-only.

This fix runs a bash loop on the device until prop sys.boot_completed
shows the reboot is complete.

Change-Id: Ie6836199e01ec1b1c87f33982bc5d960ad7ba86d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/623676
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>