Remove unused and unassigned `testInput` uniform.
To my surprise, SkRuntimeEffect does not complain if user code
declares a uniform and never sets a corresponding value, as long
as that uniform is never used. I'm not sure if that is intentional
behavior or not, but it would probably be difficult to change now.
At any rate, arbitrary unreferenced uniform declarations don't
belong in our test code.

Change-Id: I4cff5eda0cfc3602db463c483328aa36c0c6676c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/624578
Reviewed-by: Arman Uguray <armansito@google.com>