Use skvx instead of SkDVector in Path/PathBuilder
We would like to remove the dependency of SkPath on SkPathops.

Remaining dependencies:
 - SkGeometry.cpp
 - SkGlyph.cpp

While working on this, I broke something and the Paths tests
started to fail. I found those hard to understand and debug,
so I ported some of the failing ones over to PathBuilder
(which does not have as comprehensive of tests). I figured out
what I was doing wrong and what those failing tests were for.
See the additional documentation in SkPath.cpp and SkPathBuilder.cpp

This adds some new static methods to skvx, along with unit tests
for them.

Because I was comparing functionality between skvx and SkDVector,
I added some tests to the latter for an apples to apple comparison.

Change-Id: I8aa43ed021949ccf1015b2dedf26a52b59d15860
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/621198
Reviewed-by: Michael Ludwig <michaelludwig@google.com>