[graphite] Move static buffer creation to Context initialization
This moves DrawBufferManager.h to BufferManager.h and adds a related
StaticBufferManager class that's used during Context creation to
collect static vertex/index data, pack it into tight buffers (by type)
and issue copy buffer-to-buffer commands to the queue manager.

The tessellating and analytic-rrect RenderSteps are updated to use this
to create their static buffers once and then store the BindBufferInfo
directly instead of looking up in an unordered_map for each draw call.
At the moment this increases vertex usage for the tessellating renderers
since each variation no longer shares the data. A follow-up CL will
update StaticBufferManager to track this.

Updates CopyBufferToBufferTask to record source and dest offsets as
well as a size so that it can be used to copy sub ranges between
buffers.

Updates GlobalCache to be able to hold persistent refs to arbitrary
Resources so that callers can ensure a GPU resource lives as long
as the Context (currently there's no way or need to remove these refs).

Originally, SharedContext created the RendererProvider at the same
time as its other members. Now RendererProvider requires a
StaticBufferManager to create some RenderSteps, and that requires an
active ResourceProvider and QueueManager, both of which cannot happen
until subclass of SharedContext has completed its constructor. To
get around this initialization loop, a Context::initialize() function
is added that's called by Context::MakeFoo() to create the
RendererProvider and submit the GPU copy work. SharedContext adds
a setRendererProvider function that only Context can call to take
ownership of the RendererProvider if it was successful.

Bug: skia:13059
Change-Id: I941d9fad86556609f082cd16fff38a2572a42dc5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/621316
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>