Handle dropping a glyph for paths and drawables
Allow the prepareForPathDrawing and prepareForDrawableDrawing to drop
glyphs that are empty. This normally would not be a problem because
the draw mask case above would drop the glyph before doing drawable
(or path) processing. But when drawing in perspective, the draw mask
case is never called, so drawables (and paths) need to drop the empty
glyphs. Otherwise, the serialized slugs will have drawables with
nullptrs which is a problem drawing after deserialization.

Specifically, the assert in
SkScalerCache.cpp::227
fires.

Change-Id: Icddfcbe1e214c53119326dc8af65bd2d6a2480f7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/626658
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Herb Derby <herb@google.com>