Implement LValue::push in RP codegen.
Previously, we could store into an lvalue, but we didn't have a way
to push an lvalue onto the stack. Instead, we were using
pushExpression() to write the value onto the stack, and LValue::Make
to assign back into the expression.

This approach was fine for the simple lvalues we actually support
today (`myVar` and `myVar.yx`). However, it will be a problem for
more complex lvalues that contain side effects in inner
subexpressions. Arrays in particular are troublesome, e.g.
`myArray[GetArrayIndex()] *= 2` is legal in ES3. Under the previous
approach, we would need to evaluate GetArrayIndex() once to push the
left-expression onto the stack, and then evaluate it a second time
when assigning back into myArray.

Binary expressions which assign into their left-expressions now use
LValues exclusively to deal with the left side; the value is pushed
via LValue::push and written-back via LValue::store.

Change-Id: Ie50246365ba90a0309dd126498c1b2762c09ccab
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/626660
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>