Add JpegSegmentScan utility
Add the utility class JpegSegmentScan which will scan through the
segments of a Jpeg file (skipping over the entropy-encoded data
between the segments).

This utility can early-out on the first StartOfScan marker (which is
what is meant by "reading the header" in jpeglib), or it can read
through multiple StartOfImage/EndOfImage pairs.

This will be used for multiple gainmap image representations.
* For Multi Picture Format (MPF) based images, such as the ones being
  produced by shipping phones, this is needed to compute the location
  within the SkStream of the auxiliary images (this will be clear in
  the next patch).
* For JpegR images, this is needed to parse Jpeg images made of
  concatenated StartOfImage/EndOfImage pairs.

Add unit tests for this behavior.

This functionality is similar to the functionality provided by
image_io::JpegInfo.

Bug:skia:14031
Change-Id: I836de08cecfee0d5c33e1c8f18d17ae57b1bc493
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/625920
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>