SkAndroidCodec: Add getGainmapInfo
Add the private structure SkGainmapInfo, which contains the parameters
needed to render a gainmap.

Add the function SkAndroidCodec::getGainmapInfo, which will populate
those parameters, populate an SkStream for the gainmap image.

Plumb this through to SkJpegCodec and put a TODO for the next steps.

Bug: skia: 14031
Change-Id: I6957565780bcf78c64f4d096e93e59ca8d7a51e1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/625924
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>