Add gni file list for Skia codecs used by Chromium
Chromium [1] currently has a hard-coded list of sources
used for the implementation of certain codecs. This
is potentially brittle, so this change starts to introduce
some gni lists that could be used.


It is a little harder to use gn file lists like this
for Skia because GN complains if the same source file
is included multiple times (but Bazel, the source of
truth of these .gni lists, does not). I made two file
lists to slightly reduce the hard coded files in
skia/BUILD.gn as long as I was here.

In order to do this, I had to relax a constraint on
the gni_exporter tool. It would fail if the same file
was included more than once; I changed this to remove
duplicates. I also changed the file lists to be sorted
in a non-case-sensitive manner, as this is easier and
hopefully reduces churn when GN decides to (arbitrarily)
re-sort the list of files.

[1] https://source.chromium.org/chromium/chromium/src/+/main:skia/BUILD.gn;l=338;drc=f62bcc24c9aaff6ea3db5c870d813fb9c6449f04

Change-Id: Id25a7e9a018a5e2713fa91e3e0611515533afce9
Bug: skia:12541, skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/627989
Reviewed-by: Chris Mumford <cmumford@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>