Add partial RP codegen support for constant indices.
This change only allows for indexing into variables and uniforms
(things we support in LValue), and only by a literal value. This
is enough to let us say `myArray[2]` or `myMatrix[0]`. There is
more to be done for sure, but it's a start.

Change-Id: Ifb452bd2f5aa0e5b6321dccb156ab333e40c4a42
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/628476
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>