Fix assertion in SkTHashMap::find()
SkTHashMap::find() generally assumes that the collection's capacity will
exceed its element count. It expects that the "not found" case will
typically return after visiting an empty slot due to the excess
capacity. If it exhausts all the slots and does not find a match, it
asserts that the capacity must be 0.

When the collection is initialized with an initializer list, it's
possible for the count:capacity ratio to be less than 4. In particular,
when the initializer list is initialized with a single element, the
capacity is initialized to exactly 1, s.t. count == capacity. In this
state, any search for an element that's not contained in the collection
triggers the aforementioned assertion.

This assertion is now changed to check that the condition only happens
if the element count exactly matches the capacity, s.t. all slots are
non-empty AND they do not contain the requested element.

Change-Id: Iaeaec940409139096247f4c435087ef4aab78927
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/629016
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>