Add transpose support to the RP builder.
There isn't enough space in an Instruction to hold the full offset
array, so the Builder's transpose call takes the CxR of the matrix
and generates the transpose offset list itself. This is fine, but
does limit our uses of transpose to matrix-related ops. We can
always revisit this later if we have a clever idea for transpose
in the future.

Change-Id: Ia5550c18004db7802a083be571d2e8ec2afe95d6
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/628983
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>