Migrate SkRasterPipeline ops into an enum class.
In a followup CL, this will allow the op list to be moved into
a separate header entirely instead of living inside of
SkRasterPipeline.

Previously, the ops lived in `enum Stage` within SkRasterPipeline.
This led to a few confusing bits:
- The name "Stage" was highly overloaded. `SkRasterPipeline::Stage`
  was an enum of ops. `SkRasterPipelineStage` was a struct, holding
  a function pointer and context pointer.
- When we talked about "stages in a raster pipeline," we generally
  meant a full `SkRasterPipelineStage` (complete with context), but
  not always! I've tried to disambiguate between the two cases by
  changing things to read "op" where I noticed it, but I'm sure I
  didn't catch every last one.
- Ops were referred to via `SkRasterPipeline::opname` and
  `SkRasterPipeline::Stage::opname` interchangeably.
- Sometimes the enums were used as ints directly, e.g. to index
  into an array
- The Raster Pipeline code generator needed to #include all of
  SkRasterPipeline simply to name an op (although it will never
  invoke a raster-pipeline program).

This CL moves the enum outside of the SkRasterPipeline class,
renames it to `SkRasterPipelineOp`, and converts it to an `enum
class`.

Change-Id: I63f8543d5f69c43e9d6ea8fdb6a5a77d43f1ba75
Bug: skia:14042
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/629956
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>