bazel: Add script to obtain the MacSDK symlink directory.
Add a utility script to return the MacSDK directory created within
the clang_mac Bazel toolchain by `download_mac_toolchain.bzl`.
This directory contains symbolic links to the installed SDK used
by XCode.

This is intended for developer use as Bazel reserves the right
to change the output directory layout(1). This is being added because
some developers have XCode installed in a non-standard location
and having a script to identify where the symlinks should be is
helpful for troubleshooting any problems that may arise.

(1) https://bazel.build/remote/output-directories

Bug: skia:14016
Change-Id: I754f82f5ef1dd4bcd08cc8b1d9497e2888904541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/630399
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>