Coalesce adjacent calls to push_slots.
The `push_slots` op will now detect pushes of adjacent slot ranges
and combine them into a single push. This will allow us to do more
work in a single op.

This will cause multiple variables to be pushed in a single op if
their slots are adjacent, which can make the dumps read a little
less cleanly since we only print the name of the first variable
associated with the slot range. (e.g. in ArrayFollowedByScalar, we
push three scalars from an array, and a separate scalar, in a
single push, but it's listed as `rgb[0](0..3)`.)

Change-Id: I4537e342a9646d772c1d200560f0a37186453b99
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/630916
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>