[graphite] Enable AnalyticRRectRenderStep for fills
Fills for rects and round rects are mostly correct with the landed
state of AnalyticRRectRenderStep, so this enables it by default so
that it can start executing on the bots. I noticed when testing that
there are some correctness issues with the interior of rrects, but
those happened to be fixed in the process of
https://skia-review.googlesource.com/c/skia/+/631399 so I don't think
that's worth blocking on in order to get perf numbers.

Change-Id: I6e4e114eefc38ed8781806c3bde7400b2a0e1d1b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631403
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>