[graphite] Support elliptical corners for hairlines in AnalyticRRectRenderStep
The AnalyticRRectRenderStep requires circular corners for strokes so that the implicit function for the inset and the outset by the stroke half-width remains a circle. Fills can be elliptical because there's no local shift that needs to be applied. Hairlines also fall into this category, although they had been grouped with the stroke instancing since they produce bidirectional coverage.

This CL updates the Device Renderer selection logic and the instance
uploading/VS data extraction logic to signal hairlines as strokes but
in such a way that the GPU can reconstruct 8 elliptical radii instead
of assuming they are circular and only using 4 radii. It's able to do
this with a previously unused channel and the fact that join type and
stroke radius don't matter for hairlines.

Bug: b/260713170
Change-Id: Ie752bc1b96c20a87bface514edfa0a08738a4e15
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631405
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>