Change SkSRGBGammaColorFilter to SkColorSpaceXformColorFilter
We're going to need color conversion filters to build the gainmap
shader (we will need convert to a linearized version of the base
image's color space).

We already have an SkColorFilter that converts between color spaces,
it just only happens to be used for linear<->sRGB encoding. Generalize
this filter.

Bug: skia:14031
Change-Id: I160014571197b1c843bda34e8895e03129e1b97c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631098
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>