[base] Move more files into base
The general answer to "What goes in base?" is "code which
would be generally useful outside of a graphics library".

As such, I'm moving several helpers we have for managing
memory, as long as the remaining files which were needed
to be in base during the proof-of-concept build [1].

Moved from include/private into include/private/base:
 - SkNoncopyable.h

Moved from src/core into src/base:
 - SkArenaAlloc.cpp
 - SkArenaAlloc.h
 - SkArenaAllocList.h
 - SkASAN.h
 - SkAutoMalloc.h
 - SkBlockAllocator.cpp
 - SkBlockAllocator.h
 - SkMathPriv.cpp
 - SkMathPriv.h"
 - SkScopeExit.h
 - SkTBlockList.h
 - SkTDArray.cpp
 - SkThreadID.cpp
 - SkTLazy.h
 - SkUTF.cpp
 - SkUTF.h

I moved SkUNREACHABLE out of SkTypes and into SkAssert.h since
it felt assert-like.

This adds GrMemoryPool.cpp/.h to the list of files for which
we enforce IWYU, because it wasn't including <memory> and
broke when I enforced IWYU on SkBlockAllocator.

Suggested review order:
 - include/private/base/BUILD.bazel
 - include/private/BUILD.bazel
 - include/BUILD.bazel
 - gn/*.gni (these are generated from the BUILD.bazel files)
 - BUILD.gn
 - public.bzl
 - include/core/SkTypes.h and include/core/SkAssert.h
 - Everything else, mostly done with find and replace and
   cleaned up with the rewrite_includes.py script.

[1] https://skia-review.googlesource.com/c/skia/+/616777

Change-Id: I8d9a7cff89cea425a0ff4e2ee7ccb15abb021217
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631422
Reviewed-by: Ben Wagner <bungeman@google.com>
Owners-Override: Kevin Lubick <kjlubick@google.com>