Use a return mask only where it is required.
Originally, RP codegen would push/pop the return mask for every
function, and mask it off every time a return statement was
encountered. This was valid, but overkill; a return mask is only
beneficial when a function has _early_ returns. Functions with a
single return, or which only have returns at the natural end of
control flow, do not benefit from it.

We already had logic to categorize functions like this, as part
of the inliner. Now, we can reuse that logic to decide whether
or not return-mask handling is necessary for a given function.

Change-Id: I360f77c92ef962ee6a09267b230c9a8e057ba8ce
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631519
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>